+++ test/At.ts
@@ -1,12 +1,10 @@
 import { atRecord } from '../src/At/Record'
 import { atSet } from '../src/At/Set'
-import { atStrMap } from '../src/At/StrMap'
 import * as assert from 'assert'
 import { none, some } from 'fp-ts/lib/Option'
-import { setoidNumber } from 'fp-ts/lib/Setoid'
+import { eqNumber } from 'fp-ts/lib/Eq'
 import * as R from 'fp-ts/lib/Record'
 import * as S from 'fp-ts/lib/Set'
-import * as SM from 'fp-ts/lib/StrMap'
 import { Iso } from '../src'
 
 describe('At', () => {
@@ -31,30 +29,9 @@ describe('At', () => {
     })
   })
 
-  describe('atStrMap', () => {
-    const map = SM.singleton('key', 'value')
-    const at = atStrMap<string>().at('key')
-
-    it('get', () => {
-      assert.deepStrictEqual(at.get(map), some('value'))
-    })
-
-    it('add', () => {
-      const newMap = at.set(some('NEW'))(map)
-
-      assert.deepStrictEqual(newMap, SM.singleton('key', 'NEW'))
-    })
-
-    it('delete', () => {
-      const newMap = at.set(none)(map)
-
-      assert(SM.isEmpty(newMap))
-    })
-  })
-
   describe('atSet', () => {
     const set = S.singleton(3)
-    const at = atSet(setoidNumber).at(3)
+    const at = atSet(eqNumber).at(3)
 
     it('get', () => {
       assert.deepStrictEqual(at.get(set), true)
@@ -74,14 +51,14 @@ describe('At', () => {
   })
 
   it('fromIso', () => {
-    const iso = new Iso<SM.StrMap<string>, SM.StrMap<number>>(s => s.map(v => +v), a => a.map(String))
-    const at = atStrMap<number>()
+    const iso = new Iso<Record<string, string>, Record<string, number>>(R.map(v => +v), R.map(String))
+    const at = atRecord<number>()
       .fromIso(iso)
       .at('a')
-    assert.deepStrictEqual(at.get(new SM.StrMap({})), none)
-    assert.deepStrictEqual(at.get(new SM.StrMap({ a: '1' })), some(1))
+    assert.deepStrictEqual(at.get({}), none)
+    assert.deepStrictEqual(at.get({ a: '1' }), some(1))
 
-    assert.deepStrictEqual(at.set(none)(new SM.StrMap({})), new SM.StrMap({}))
-    assert.deepStrictEqual(at.set(some(1))(new SM.StrMap({})), new SM.StrMap({ a: '1' }))
+    assert.deepStrictEqual(at.set(none)({}), {})
+    assert.deepStrictEqual(at.set(some(1))({}), { a: '1' })
   })
 })