+++ test/Optional.ts
@@ -1,22 +1,32 @@
 import * as assert from 'assert'
 import { Optional, Lens } from '../src'
-import { none, some, Option } from 'fp-ts/lib/Option'
+import * as O from 'fp-ts/lib/Option'
 import { identity } from 'fp-ts/lib/function'
+import { pipe } from 'fp-ts/lib/pipeable'
 
 interface A {
-  a: Option<number>
+  a: O.Option<number>
 }
 
-const optional = new Optional<A, number>(s => s.a, a => s => s.a.fold<A>(s, () => ({ ...s, a: some(a) })))
+const optional = new Optional<A, number>(
+  s => s.a,
+  a => s => {
+    if (O.isNone(s.a)) {
+      return s
+    } else {
+      return { ...s, a: O.some(a) }
+    }
+  }
+)
 
 describe('Optional', () => {
   it('getOption', () => {
-    assert.deepStrictEqual(optional.getOption({ a: none }), none)
-    assert.deepStrictEqual(optional.getOption({ a: some(1) }), some(1))
+    assert.deepStrictEqual(optional.getOption({ a: O.none }), O.none)
+    assert.deepStrictEqual(optional.getOption({ a: O.some(1) }), O.some(1))
   })
 
   it('set', () => {
-    assert.deepStrictEqual(optional.set(2)({ a: some(1) }).a, some(2))
+    assert.deepStrictEqual(optional.set(2)({ a: O.some(1) }).a, O.some(2))
   })
 
   it('fromNullableProp', () => {
@@ -33,14 +43,7 @@ describe('Optional', () => {
       info?: Info
     }
 
-    const info1 = Optional.fromNullableProp<Response, Info, 'info'>('info')
-    const employment1 = Optional.fromNullableProp<Info, Employment, 'employment'>('employment')
-    const phone1 = Optional.fromNullableProp<Employment, Phone, 'phone'>('phone')
     const phoneNumber = Lens.fromProp<Phone>()('number')
-    const numberFromResponse1 = info1
-      .compose(employment1)
-      .compose(phone1)
-      .composeLens(phoneNumber)
 
     const response1: Response = {
       info: {
@@ -57,9 +60,17 @@ describe('Optional', () => {
       }
     }
 
-    assert.deepStrictEqual(numberFromResponse1.getOption(response1), some('555-1234'))
-    assert.deepStrictEqual(numberFromResponse1.getOption(response2), none)
-    assert.deepStrictEqual(numberFromResponse1.set('b')(response1), {
+    const info = Optional.fromNullableProp<Response>()('info')
+    const employment = Optional.fromNullableProp<Info>()('employment')
+    const phone = Optional.fromNullableProp<Employment>()('phone')
+    const numberFromResponse = info
+      .compose(employment)
+      .compose(phone)
+      .composeLens(phoneNumber)
+
+    assert.deepStrictEqual(numberFromResponse.getOption(response1), O.some('555-1234'))
+    assert.deepStrictEqual(numberFromResponse.getOption(response2), O.none)
+    assert.deepStrictEqual(numberFromResponse.set('b')(response1), {
       info: {
         employment: {
           phone: {
@@ -68,26 +79,10 @@ describe('Optional', () => {
         }
       }
     })
-    assert.deepStrictEqual(numberFromResponse1.set('b')(response2), response2)
-    assert.strictEqual(numberFromResponse1.set('555-1234')(response1), response1)
-    assert.strictEqual(numberFromResponse1.modify(identity)(response1), response1)
-    assert.strictEqual(numberFromResponse1.modify(identity)(response2), response2)
-
-    const info2 = Optional.fromNullableProp<Response>()('info')
-    const employment2 = Optional.fromNullableProp<Info>()('employment')
-    const phone2 = Optional.fromNullableProp<Employment>()('phone')
-    const numberFromResponse2 = info2
-      .compose(employment2)
-      .compose(phone2)
-      .composeLens(phoneNumber)
-
-    assert.deepStrictEqual(numberFromResponse2.getOption(response1), numberFromResponse1.getOption(response1))
-    assert.deepStrictEqual(numberFromResponse2.getOption(response2), numberFromResponse1.getOption(response2))
-    assert.deepStrictEqual(numberFromResponse2.set('b')(response1), numberFromResponse1.set('b')(response1))
-    assert.deepStrictEqual(numberFromResponse2.set('b')(response2), numberFromResponse1.set('b')(response2))
-    assert.strictEqual(numberFromResponse2.set('555-1234')(response1), response1)
-    assert.strictEqual(numberFromResponse2.modify(identity)(response1), response1)
-    assert.strictEqual(numberFromResponse2.modify(identity)(response2), response2)
+    assert.deepStrictEqual(numberFromResponse.set('b')(response2), response2)
+    assert.strictEqual(numberFromResponse.set('555-1234')(response1), response1)
+    assert.strictEqual(numberFromResponse.modify(identity)(response1), response1)
+    assert.strictEqual(numberFromResponse.modify(identity)(response2), response2)
   })
 
   it('fromOptionProp', () => {
@@ -95,60 +90,50 @@ describe('Optional', () => {
       number: string
     }
     interface Employment {
-      phone: Option<Phone>
+      phone: O.Option<Phone>
     }
     interface Info {
-      employment: Option<Employment>
+      employment: O.Option<Employment>
     }
     interface Response {
-      info: Option<Info>
+      info: O.Option<Info>
     }
 
-    const info1 = Optional.fromOptionProp<Response>('info')
-    const employment1 = Optional.fromOptionProp<Info>('employment')
-    const phone1 = Optional.fromOptionProp<Employment>('phone')
     const phoneNumber = Lens.fromProp<Phone>()('number')
-    const numberFromResponse1 = info1
-      .compose(employment1)
-      .compose(phone1)
-      .composeLens(phoneNumber)
 
     const response1: Response = {
-      info: some({
-        employment: some({
-          phone: some({
+      info: O.some({
+        employment: O.some({
+          phone: O.some({
             number: '555-1234'
           })
         })
       })
     }
     const response2: Response = {
-      info: some({
-        employment: none
+      info: O.some({
+        employment: O.none
       })
     }
 
-    assert.deepStrictEqual(numberFromResponse1.getOption(response1), some('555-1234'))
-    assert.deepStrictEqual(numberFromResponse1.getOption(response2), none)
-
-    const info2 = Optional.fromOptionProp<Response>()('info')
-    const employment2 = Optional.fromOptionProp<Info>()('employment')
-    const phone2 = Optional.fromOptionProp<Employment>()('phone')
-    const numberFromResponse2 = info2
-      .compose(employment2)
-      .compose(phone2)
+    const info = Optional.fromOptionProp<Response>()('info')
+    const employment = Optional.fromOptionProp<Info>()('employment')
+    const phone = Optional.fromOptionProp<Employment>()('phone')
+    const numberFromResponse = info
+      .compose(employment)
+      .compose(phone)
       .composeLens(phoneNumber)
 
-    assert.deepStrictEqual(numberFromResponse2.getOption(response1), some('555-1234'))
-    assert.deepStrictEqual(numberFromResponse2.getOption(response2), none)
+    assert.deepStrictEqual(numberFromResponse.getOption(response1), O.some('555-1234'))
+    assert.deepStrictEqual(numberFromResponse.getOption(response2), O.none)
 
     // Check the laws
-    const opt = numberFromResponse1
+    const opt = numberFromResponse
     // Law 1
     assert.deepStrictEqual(opt.set('555-4321')(response1), {
-      info: some({
-        employment: some({
-          phone: some({
+      info: O.some({
+        employment: O.some({
+          phone: O.some({
             number: '555-4321'
           })
         })
@@ -158,11 +143,17 @@ describe('Optional', () => {
     // Law 2
     assert.deepStrictEqual(
       opt.getOption(opt.set('555-4321')(response1)),
-      opt.getOption(response1).map(() => '555-4321')
+      pipe(
+        opt.getOption(response1),
+        O.map(() => '555-4321')
+      )
     )
     assert.deepStrictEqual(
       opt.getOption(opt.set('555-4321')(response2)),
-      opt.getOption(response2).map(() => '555-4321')
+      pipe(
+        opt.getOption(response2),
+        O.map(() => '555-4321')
+      )
     )
     // law 3
     assert.deepStrictEqual(opt.set('555-4321')(opt.set('555-4321')(response1)), opt.set('555-4321')(response1))
@@ -171,48 +162,50 @@ describe('Optional', () => {
 
   it('modify', () => {
     const double = (n: number): number => n * 2
-    assert.deepStrictEqual(optional.modify(double)({ a: some(2) }), { a: some(4) })
-    assert.deepStrictEqual(optional.modify(double)({ a: none }), { a: none })
+    assert.deepStrictEqual(optional.modify(double)({ a: O.some(2) }), { a: O.some(4) })
+    assert.deepStrictEqual(optional.modify(double)({ a: O.none }), { a: O.none })
   })
 
   it('compose', () => {
     interface Phone {
-      number: Option<string>
+      number: O.Option<string>
     }
     interface Employment {
-      phone: Option<Phone>
+      phone: O.Option<Phone>
     }
     const phone = Optional.fromOptionProp<Employment>()('phone')
     const phoneNumber = Optional.fromOptionProp<Phone>()('number')
     const composition1 = phone.compose(phoneNumber)
     const composition2 = phone.composeOptional(phoneNumber)
-    assert.deepStrictEqual(composition1.getOption({ phone: none }), none)
-    assert.deepStrictEqual(composition1.getOption({ phone: some({ number: none }) }), none)
-    assert.deepStrictEqual(composition1.getOption({ phone: some({ number: some('a') }) }), some('a'))
-    assert.deepStrictEqual(composition1.set('a')({ phone: none }), { phone: none })
-    assert.deepStrictEqual(composition1.set('a')({ phone: some({ number: none }) }), { phone: some({ number: none }) })
-    assert.deepStrictEqual(composition1.set('a')({ phone: some({ number: some('b') }) }), {
-      phone: some({ number: some('a') })
+    assert.deepStrictEqual(composition1.getOption({ phone: O.none }), O.none)
+    assert.deepStrictEqual(composition1.getOption({ phone: O.some({ number: O.none }) }), O.none)
+    assert.deepStrictEqual(composition1.getOption({ phone: O.some({ number: O.some('a') }) }), O.some('a'))
+    assert.deepStrictEqual(composition1.set('a')({ phone: O.none }), { phone: O.none })
+    assert.deepStrictEqual(composition1.set('a')({ phone: O.some({ number: O.none }) }), {
+      phone: O.some({ number: O.none })
+    })
+    assert.deepStrictEqual(composition1.set('a')({ phone: O.some({ number: O.some('b') }) }), {
+      phone: O.some({ number: O.some('a') })
     })
 
-    assert.deepStrictEqual(composition2.getOption({ phone: none }), composition1.getOption({ phone: none }))
+    assert.deepStrictEqual(composition2.getOption({ phone: O.none }), composition1.getOption({ phone: O.none }))
     assert.deepStrictEqual(
-      composition2.getOption({ phone: some({ number: none }) }),
-      composition1.getOption({ phone: some({ number: none }) })
+      composition2.getOption({ phone: O.some({ number: O.none }) }),
+      composition1.getOption({ phone: O.some({ number: O.none }) })
     )
     assert.deepStrictEqual(
-      composition2.getOption({ phone: some({ number: some('a') }) }),
-      composition1.getOption({ phone: some({ number: some('a') }) })
+      composition2.getOption({ phone: O.some({ number: O.some('a') }) }),
+      composition1.getOption({ phone: O.some({ number: O.some('a') }) })
     )
 
-    assert.deepStrictEqual(composition2.set('a')({ phone: none }), composition1.set('a')({ phone: none }))
+    assert.deepStrictEqual(composition2.set('a')({ phone: O.none }), composition1.set('a')({ phone: O.none }))
     assert.deepStrictEqual(
-      composition2.set('a')({ phone: some({ number: none }) }),
-      composition1.set('a')({ phone: some({ number: none }) })
+      composition2.set('a')({ phone: O.some({ number: O.none }) }),
+      composition1.set('a')({ phone: O.some({ number: O.none }) })
     )
     assert.deepStrictEqual(
-      composition2.set('a')({ phone: some({ number: some('b') }) }),
-      composition1.set('a')({ phone: some({ number: some('b') }) })
+      composition2.set('a')({ phone: O.some({ number: O.some('b') }) }),
+      composition1.set('a')({ phone: O.some({ number: O.some('b') }) })
     )
   })
 })